mixed Simple.prototype into CellEditor.prototype; delete Simple.js
