+++ src/cellEditors/Color.js
@@ -1,11 +1,11 @@
 'use strict';
 
-var Simple = require('./Simple');
+var CellEditor = require('./CellEditor');
 
 /**
  * @constructor
  */
-var Color = Simple.extend('Color', {
+var Color = CellEditor.extend('Color', {
 
     template: function() {
         /*
