+++ src/cellEditors/Textfield.js
@@ -1,11 +1,11 @@
 'use strict';
 
-var Simple = require('./Simple');
+var CellEditor = require('./CellEditor.js');
 
 /**
  * @constructor
  */
-var Textfield = Simple.extend('Textfield', {
+var Textfield = CellEditor.extend('Textfield', {
 
     template: function() {
         /*
